backend completo

+++ backend/services/categoria_service.go
@@ -11,13 +11,24 @@ import (
 
 type CategoriaRepository interface {
 
-	GetAll() ([]models.Categoria, error)
+
+	GetAll(page int, limit int, busca string) ([]models.Categoria, int, error)
+
+
+	GetByID(id string) (models.Categoria, error)
+
 
 	Create(categoria *models.Categoria) error
 
+
 	Update(id string, categoria models.Categoria) error
 
+
 	Delete(id string) error
+
+
+	ExisteProdutoVinculado(id string) (bool, error)
+
 }
 
 
@@ -25,6 +36,7 @@ type CategoriaRepository interface {
 type CategoriaService struct {
 
 	repository CategoriaRepository
+
 }
 
 
@@ -32,16 +44,31 @@ type CategoriaService struct {
 func NewCategoriaService() *CategoriaService {
 
 	return &CategoriaService{
+
 		repository: repositories.NewCategoriaRepository(),
+
 	}
+
 }
 
 
 
-func (s *CategoriaService) GetAll() ([]models.Categoria, error) {
 
-	return s.repository.GetAll()
+func (s *CategoriaService) GetAll(page int, limit int, busca string) ([]models.Categoria, int, error) {
+
+	return s.repository.GetAll(page, limit, busca)
+
 }
+
+
+
+
+func (s *CategoriaService) GetByID(id string) (models.Categoria, error) {
+
+	return s.repository.GetByID(id)
+
+}
+
 
 
 
@@ -51,11 +78,16 @@ func (s *CategoriaService) Create(categoria *models.Categoria) error {
 	if !utils.ValidarTexto(categoria.Nome) {
 
 		return fmt.Errorf("nome da categoria é obrigatório")
+
 	}
 
 
+
 	return s.repository.Create(categoria)
+
 }
+
+
 
 
 
@@ -65,15 +97,45 @@ func (s *CategoriaService) Update(id string, categoria models.Categoria) error {
 	if !utils.ValidarTexto(categoria.Nome) {
 
 		return fmt.Errorf("nome da categoria é obrigatório")
+
 	}
 
 
+
 	return s.repository.Update(id, categoria)
+
 }
+
+
 
 
 
 func (s *CategoriaService) Delete(id string) error {
 
+
+	existeProduto, err := s.repository.ExisteProdutoVinculado(id)
+
+
+	if err != nil {
+
+		return err
+
+	}
+
+
+
+	if existeProduto {
+
+
+		return fmt.Errorf(
+			"não é possível excluir a categoria pois existem produtos vinculados",
+		)
+
+	}
+
+
+
+
 	return s.repository.Delete(id)
+
 }
